chore: extract components rom theme-token
